omdøp FlexSykmelding -> Sykmelding
